Refactor backButton listener to hook
