+++ src/app/components/Ionic/IonicProvider.tsx
@@ -1,32 +1,20 @@
 import { createContext, FC, PropsWithChildren, useEffect, useRef } from 'react'
-import { useNavigate } from 'react-router-dom'
 import { Capacitor } from '@capacitor/core'
 import { App } from '@capacitor/app'
 import { useDispatch } from 'react-redux'
 import { persistActions } from '../../state/persist'
 import { deltaMsToLockProfile } from '../../../ionicConfig'
+import { useIonicBackButtonListener } from './hooks/useIonicBackButtonListener'
 
 const IonicContext = createContext<undefined>(undefined)
 
 const IonicContextProvider: FC<PropsWithChildren> = ({ children }) => {
   const dispatch = useDispatch()
-  const navigate = useNavigate()
   const lockTimestamp = useRef<number>()
 
-  useEffect(() => {
-    /**
-     * The back button refers to the physical back button on an Android device and should not be confused
-     * with either the browser back button or ion-back-button.
-     */
-    const backButtonListenerHandle = App.addListener('backButton', ({ canGoBack }) => {
-      if (!canGoBack) {
-        App.exitApp()
-      } else {
-        navigate(-1)
-      }
-    })
+  useIonicBackButtonListener()
 
-    // TODO: appStateChange triggers on the web platform as well, using visibilitychange listener, consider reusing the code(downside @capacitor/app dependency on web & extension)
+  useEffect(() => {
     const appStateChangeListenerHandle = App.addListener('appStateChange', ({ isActive }) => {
       const shouldLock = lockTimestamp.current && Date.now() - lockTimestamp.current > deltaMsToLockProfile
       if (isActive && shouldLock) {
@@ -41,10 +29,9 @@ const IonicContextProvider: FC<PropsWithChildren> = ({ children }) => {
     })
 
     return () => {
-      backButtonListenerHandle.remove()
       appStateChangeListenerHandle.remove()
     }
-  }, [dispatch, navigate])
+  }, [dispatch])
 
   return <IonicContext.Provider value={undefined}>{children}</IonicContext.Provider>
 }
